intel/whois-testing/requirements.txt
python-whois==0.9.5
requests==2.32.5
pytest==9.0.3
